feat(components): Enhance `ProfileImageToken` previews
This commit enhances the preview for the `ProfileImageToken` component to display a more comprehensive set of variations.

The preview now utilizes `FlowRow` to showcase all possible `AvatarBackground` colors for the profile image. It also demonstrates the component in both its `hero` and default states, with and without a specific `Avatar` set, providing a clearer visual reference for different use cases.