feat(logging): Add error logging to `IosProfileStore`
This commit introduces error logging to the `IosProfileStore` on the iOS platform.

A `Logger` instance has been added to the class. If an exception occurs while decoding the user's profile from `NSUserDefaults`, the error message and exception will now be logged. This improves observability and makes it easier to diagnose issues related to corrupt or invalid stored profile data.